Added custom theme hook + updated dark color scheme
