+++ src/components/Home.js
@@ -1,17 +1,17 @@
-import React from 'react'
+import React from "react";
 
-import InfoPane from './InfoPane';
-import Settings from './Settings';
+import InfoPane from "./InfoPane";
+import Settings from "./Settings";
 
 const Home = () => {
-    return (
-        <div className="Home">
-            <div className="container">
-                <InfoPane />
-                <Settings />
-            </div>
-        </div>
-    )
-}
+  return (
+    <div className="Home">
+      <div className="container">
+        <InfoPane />
+        <Settings />
+      </div>
+    </div>
+  );
+};
 
-export default Home
+export default Home;
